miepgui:  Added menu for jpg and mp4 export. writeJPG and wirte MP4 now take miep settings as second argument
readBBX: Nochmals? is now even mor annoying. It does not allow for nan values anymore.

writeJPG: preliminary working version of jpg export. Now takes miep settings as second argument instead of export folder. if no miep settings are given, it searches the standard paths and defaults to standard color maps if no miepsettings are found or the imported settings are not an object.

writeMP4: pretty much the same as jpg export but for movies